Remove B/C HTML structure and improve mobile view
When I first converted the blog to GitHub pages we were still
supporting versions of IE that didn't understand HTML5 properly
so the structure had <div> elements to provide backwards
compatibility for those browsers. We're now four years on and
we no longer support those browsers so tidy up the HTML and CSS.

Also as part of this the mobile view has been improved by using
breakpoints to adjust font-sizing.